movie controller added
